+++ src/Components/Config.ts
@@ -1,22 +1,7 @@
 const ENV_DEVELOPMENT = 'development'
-const ENV_STAGING = 'staging'
 const ENV_PRODUCTION = 'production'
 
 const CURRENT_ENV = process.env.REACT_APP_ENV || ENV_PRODUCTION
-
-const STAGING = {
-    CHAIN_TAG: '0x27',
-    THOR: 'https://thor-staging.decent.bet',
-    GETH: 'wss://geth-staging.decent.bet/ws',
-    GETH_RPC: 'https://rinkeby.infura.io/v3/d6d6d340ec174822ab1892bf3db85bda',
-    VEFORGE: 'https://testnet.veforge.com',
-    ETHERSCAN: 'https://rinkeby.etherscan.io',
-    DEPOSIT_ADDR: '0x4b484ec17b4b5a16a5adfa8e4b92f74050af24ca',
-    V1_TOKEN_ADDR: '0xa48d81beab4a5d68a5623d352889d79a8ec205c9',
-    V2_TOKEN_ADDR: '0xf1d5080c0c96a80325148378df1ab7b9c6ceed5d',
-    VET_TOKEN_ADDR: '0x510fCddC9424B1bBb328A574f45BfDdB130e1f03',
-    V1_UPGRADE_AGENT_ADDR: '0x2139ECD8246594Ef21267EafDAe924130f967a3E',
-}
 
 const DEVELOPMENT = {
     CHAIN_TAG: '0x27',
@@ -36,7 +21,7 @@ const PRODUCTION = {
     CHAIN_TAG: '0xc7',
     THOR: 'https://thor.decent.bet',
     GETH: 'wss://geth.decent.bet/ws',
-    GETH_RPC: 'https://mainnet.infura.io',
+    GETH_RPC: 'https://geth.decent.bet/http',
     VEFORGE: 'https://explore.veforge.com',
     ETHERSCAN: 'https://etherscan.io',
     DEPOSIT_ADDR: '0x3f1e4ef0b246eb95fc73f18f6613b799811a739b',
@@ -48,8 +33,7 @@ const PRODUCTION = {
 
 const Configs: any = {
     [ENV_PRODUCTION]: PRODUCTION,
-    [ENV_DEVELOPMENT]: DEVELOPMENT,
-    [ENV_STAGING]: STAGING
+    [ENV_DEVELOPMENT]: DEVELOPMENT
 }
 export class Config {
     static get chainTag() {