+++ src/Components/Web3/ContractHelper.ts
@@ -7,6 +7,8 @@ export default class ContractHelper {
     protected web3: Web3;
     private v1TokenContract: DBETV1TokenContract;
     private v2TokenContract: DBETV2TokenContract;
+    private v1TokenContract_Http: DBETV1TokenContract;
+    private v2TokenContract_Http: DBETV2TokenContract;
     private depositContract: DBETToVETDepositContract;
     private vetContract: DBETVETTokenContract;
     /**
@@ -19,6 +21,8 @@ export default class ContractHelper {
         // Initialize new Contracts
         this.v1TokenContract = new DBETV1TokenContract(this.web3)
         this.v2TokenContract = new DBETV2TokenContract(this.web3)
+        this.v1TokenContract_Http = new DBETV1TokenContract((window as any).web3Http)
+        this.v2TokenContract_Http = new DBETV2TokenContract((window as any).web3Http)
         this.depositContract = new DBETToVETDepositContract(this.web3, thor)
         this.vetContract = new DBETVETTokenContract(this.web3, thor)
 
@@ -31,6 +35,15 @@ export default class ContractHelper {
     get V2Token() {
         return this.v2TokenContract
     }
+
+    get V1TokenHttp() {
+        return this.v1TokenContract_Http
+    }
+
+    get V2TokenHttp() {
+        return this.v2TokenContract_Http
+    }
+
 
     get DepositToVET() {
         return this.depositContract